Removed a lot of superfluous code
Ok, so what happened here?

* I realised that I was doing extra randomisation work: all particles started in a random place, so why pick a random next particle? Just pick the first one that hasn't been followed yet (i.e. use nextTargetIndex). That removed the need for friendFinder
* Hold on, if I use an askedIndex that lags behind nextTargetIndex, then whether or not a particle should be drawn (isDrifting), is being followed (isFollowed) , befriended or unavailable (isNotAvailable) becomes implicit to the structure of the array + those two indices + the state of isBefriended. So I can remove isDrifting, isFollowed and isNotAvailable too!

So this simplified the code A LOT. We're almost back to the beginning, but without losing most of the added functionality.

The downside is that you have to be aware of the information implicit in the data structure to follow what is going on, so it get's a bit obfuscated. On the flipside, it *almost* runs smoothly on my mobile browser now!